+++ src/search-strategies/index.ts
@@ -2,3 +2,4 @@ export {
   LoopedIndexOfAnchoredSearchStrategy as StringAnchorSearchStrategy
 } from "./looped-indexOf-anchored/index.js";
 export * from "./regex/index.js";
+export * from "./balanced-pair/index.js";